@@ -1,5 +1,5 @@
 {
-  "name": "carolineleger",
+  "name": "author",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -8,20 +8,23 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "core-js": "^3.22.5",
-    "node-sass": "^4.14.1",
-    "sass-loader": "^10.2.1",
-    "vue": "^2.6.14"
+    "@fortawesome/fontawesome-free": "^6.4.2",
+    "@mdi/font": "^7.2.96",
+    "core-js": "^3.8.3",
+    "sass": "^1.66.1",
+    "sass-loader": "^13.3.2",
+    "vue": "^3.2.13",
+    "vuetify": "^3.3.15",
+    "wowjs": "^1.1.3"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "^4.5.17",
-    "@vue/cli-plugin-eslint": "^4.5.17",
-    "@vue/cli-service": "^4.5.17",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "vue-template-compiler": "^2.6.14",
-    "wowjs": "^1.1.3"
+    "@babel/core": "^7.12.16",
+    "@babel/eslint-parser": "^7.12.16",
+    "@vue/cli-plugin-babel": "~5.0.0",
+    "@vue/cli-plugin-eslint": "~5.0.0",
+    "@vue/cli-service": "~5.0.0",
+    "eslint": "^7.32.0",
+    "eslint-plugin-vue": "^8.0.3"
   },
   "eslintConfig": {
     "root": true,
@@ -29,17 +32,18 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
+      "plugin:vue/vue3-essential",
       "eslint:recommended"
     ],
     "parserOptions": {
-      "parser": "babel-eslint"
+      "parser": "@babel/eslint-parser"
     },
     "rules": {}
   },
   "browserslist": [
     "> 1%",
     "last 2 versions",
-    "not dead"
+    "not dead",
+    "not ie 11"
   ]
 }
