@@ -1,8 +1,19 @@
-import Vue from 'vue'
-import App from './App.vue'
+import "@mdi/font/css/materialdesignicons.css";
+import "@fortawesome/fontawesome-free/css/all.css";
+import { createApp } from "vue";
+// Vuetify
+import "vuetify/styles";
+import { createVuetify } from "vuetify";
+import * as components from "vuetify/components";
+import * as directives from "vuetify/directives";
 
-Vue.config.productionTip = false
+const vuetify = createVuetify({
+  components,
+  directives,
+});
 
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+import App from "./App.vue";
+
+const app = createApp(App).use(vuetify);
+
+app.mount("#app");
